Modifying the java.rb

+++ lib/java.rb
@@ -42,16 +42,25 @@ class Java < Code
       stdin, stdout, stderr = Open3.popen3('javac Code.java')
 
       puts stderr.gets
-      $result = `timeout 4s java Code` if File.exist?("Code.class")
 
-      if $result.chomp != tc.output
-        File.delete("Code.java") if File.exist?("Code.java")
-        File.delete("Code.class") if File.exist?("Code.class")
+      if File.exist?("Code.class")
+
+        $result = `timeout 4s java Code`
+
+        if $result.chomp != tc.output
+          File.delete("Code.java") if File.exist?("Code.java")
+          File.delete("Code.class") if File.exist?("Code.class")
+          return false
+        end
+
+      else
+
         return false
+
       end
 
-      File.delete("Code.java")
-      File.delete("Code.class")
+      File.delete("Code.java") if File.exist?("Code.java")
+      File.delete("Code.class") if File.exist?("Code.java")
 
     end
